+++ src/accounts.rs
@@ -74,9 +74,10 @@ mod tests {
             base_vault: Pubkey::new_unique(),
             quote_vault: Pubkey::new_unique(),
             base_reserve: 1_000_000_000_000_000,
-            quote_reserve: 10_000_000_000,
+            quote_reserve: 5_000_000_000,
             virtual_base_reserve: 1_000_000_000_000_000,
             virtual_quote_reserve: 20_000_000_000,
+            is_migrated: 0,
         }
     }